Feat : before create imgurls & hashtags
