+++ src/pages/_app.tsx
@@ -1,8 +1,25 @@
 import "../styles/globals.css";
 import type { AppProps } from "next/app";
+import { GetStaticProps } from "next";
 
 function MyApp({ Component, pageProps }: AppProps) {
   return <Component {...pageProps} />;
 }
+
+export const getStaticProps: GetStaticProps = async () => {
+  const REGION = process.env.AWS_REGION ? process.env.AWS_REGION : null;
+  const ACCESS_KEY = process.env.AWS_KEY ? process.env.AWS_KEY : null;
+  const SECRECT_KEY = process.env.AWS_SECRET_KEY
+    ? process.env.AWS_SECRET_KEY
+    : null;
+
+  return {
+    props: {
+      region: REGION,
+      accessKey: ACCESS_KEY,
+      secretKey: SECRECT_KEY,
+    },
+  };
+};
 
 export default MyApp;
